Move `markdown` and `documentation` into  seperate modules
