+++ cedar-language-server/src/documentation/extension/ip.rs
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-use crate::markdown::{MarkdownBuilder, ToDocumentationString};
+use super::ToDocumentationString;
+use crate::markdown::MarkdownBuilder;
 use cedar_policy_core::validator::ValidatorSchema;
 use indoc::indoc;
